style/#74_refactoring: moved card and deck to Phaser dir
