+++ src/Phaser/blackjack/BlackjackScene.ts
@@ -5,9 +5,9 @@ import Button from '../common/button'
 import Text = Phaser.GameObjects.Text
 import Zone = Phaser.GameObjects.Zone
 import { CARD_HEIGHT, CARD_WIDTH } from '@/Factories/cardFactory'
+import Card from '@/Phaser/common/CardImage'
+import Deck from '@/Phaser/common/DeckImage'
 import BlackjackPlayer from '@/model/blackjack/BlackjackPlayer'
-import Card from '@/model/common/CardImage'
-import Deck from '@/model/common/DeckImage'
 import GameResult from '@/model/common/gameResult'
 import { Result } from '@/model/common/types/game'
 import ImageUtility from '@/utility/ImageUtility'
