+++ src/Phaser/war/WarScene.ts
@@ -4,8 +4,8 @@ import Button from '../common/button'
 import Text = Phaser.GameObjects.Text
 import Zone = Phaser.GameObjects.Zone
 import { CARD_HEIGHT, CARD_WIDTH } from '@/Factories/cardFactory'
-import Card from '@/model/common/CardImage'
-import Deck from '@/model/common/DeckImage'
+import Card from '@/Phaser/common/CardImage'
+import Deck from '@/Phaser/common/DeckImage'
 import GameResult from '@/model/common/gameResult'
 import WarPlayer from '@/model/war/WarPlayer'
 import { GUTTER_SIZE, textStyle } from '@/utility/constants'
